+++ jest.setup.js
@@ -1,3 +1,4 @@
+import '@testing-library/jest-dom';
 import { configure } from '@testing-library/react';
 
 import './.jest/extend-expect';