+++ src/react/components/dateinput/XUIDateInputWIP.js
@@ -37,8 +37,10 @@ class XUIDateInputWIP extends React.Component {
       locale,
       maxDate,
       minDate,
+      nextButtonAriaLabel,
       onInputChange,
       onSelectDate, // Destructured so as not to spread.
+      prevButtonAriaLabel,
       selectedDateValue, // Destructured so as not to spread.
       selectedDateDefaultValue, // Destructured so as not to spread.
       selectDateLabel,
@@ -63,8 +65,10 @@ class XUIDateInputWIP extends React.Component {
           locale={locale}
           maxDate={maxDate}
           minDate={minDate}
+          nextButtonAriaLabel={nextButtonAriaLabel}
           onInputChange={onInputChange}
           onSelectDate={this.onSelectDate}
+          prevButtonAriaLabel={prevButtonAriaLabel}
           qaHook={qaHook && `${qaHook}-dateinput`}
           selectDateLabel={selectDateLabel}
           selectedDate={selectedDate}
@@ -103,7 +107,7 @@ XUIDateInputWIP.propTypes = {
   hintMessage: PropTypes.string,
 
   /** Input label */
-  inputLabel: PropTypes.string,
+  inputLabel: PropTypes.string.isRequired,
 
   /** Whether the input is disabled */
   isDisabled: PropTypes.bool,
@@ -111,8 +115,8 @@ XUIDateInputWIP.propTypes = {
   /** Whether the current input value is invalid */
   isInvalid: PropTypes.bool,
 
-  /** The locale of the calendar. Defaults to En */
-  locale: PropTypes.string,
+  /** The locale of the calendar. */
+  locale: PropTypes.string.isRequired,
 
   /**
    * If you want to disable every date after a given day, pass in the maximum enabled
@@ -126,11 +130,25 @@ XUIDateInputWIP.propTypes = {
    */
   minDate: PropTypes.instanceOf(Date),
 
+  /** An accessibility label for the next month button that will be used
+   * by assistive technologies.
+   *
+   * Recommended English value: *Next month*
+   */
+  nextButtonAriaLabel: PropTypes.string.isRequired,
+
   /** Callback for when the input changes  */
   onInputChange: PropTypes.func,
 
   /** Callback for when the user selects a date */
   onSelectDate: PropTypes.func,
+
+  /** An accessibility label for the previous month button that will be used
+   * by assistive technologies.
+   *
+   * Recommended English value: *Previous month*
+   */
+  prevButtonAriaLabel: PropTypes.string.isRequired,
 
   qaHook: PropTypes.string,
 
@@ -152,7 +170,6 @@ XUIDateInputWIP.propTypes = {
 XUIDateInputWIP.defaultProps = {
   closeOnSelect: true,
   displayedMonth: new Date(),
-  locale: 'en',
 };
 
 export default XUIDateInputWIP;