+++ src/react/components/dateinput/private/XUIDateInputItem.d.ts
@@ -28,7 +28,7 @@ interface Props {
   inputFieldClassName?: string;
 
   /** Input label */
-  inputLabel?: string;
+  inputLabel: string;
 
   /** Whether the input is disabled */
   isDisabled?: boolean;
@@ -36,8 +36,8 @@ interface Props {
   /** Whether the current input value is invalid */
   isInvalid?: boolean;
 
-  /** The locale of the calendar. Defaults to En */
-  locale?: string;
+  /** The locale of the calendar. */
+  locale: string;
 
   /** If you want to disable every date after a given day, pass in the maximum enabled
    * date here.  Can be used with the isDateDisabled function. */
@@ -47,11 +47,27 @@ interface Props {
    * date here.  Can be used with the isDateDisabled function. */
   minDate?: Date;
 
+  /**
+   * An accessibility label for the previous month button that will be used
+   * by assistive technologies.
+   *
+   * Recommended English value: *Previous month*
+   */
+  nextButtonAriaLabel: string;
+
   /** Callback for when the input changes  */
   onInputChange?: (event: React.MouseEvent<HTMLDivElement>) => void;
 
   /** Callback for when the user selects a date */
   onSelectDate?: (day: Date) => void;
+
+  /**
+   * An accessibility label for the previous month button that will be used
+   * by assistive technologies.
+   *
+   * Recommended English value: *Previous month*
+   */
+  prevButtonAriaLabel: string;
 
   qaHook?: string;