+++ src/react/components/dateinput/stories/stories.js
@@ -25,20 +25,26 @@ storiesWithKnobs.add('Playground', () => {
   const isDateRangeDemo = boolean('isDateRangeInput', false);
 
   let singleDateProps = {
-    firstInputLabel: 'Single date Date',
     closeOnSelect: true,
+    inputLabel: 'Single date',
+    locale: 'en',
+    nextButtonAriaLabel: 'Next month',
     onSelectDate: sampleOnSelectDateFunction,
+    prevButtonAriaLabel: 'Previous month',
     validationMessage: text('validationMessage', ''),
   };
 
   let dateRangeProps = {
+    convenienceDates: dateRangeInputConvenienceDates,
+    locale: 'en',
+    nextButtonAriaLabel: 'Next month',
+    prevButtonAriaLabel: 'Previous month',
     startDateInputConfig: {
       onInputChange: sampleOnSelectDateFunction,
     },
     endDateInputConfig: {
       onInputChange: sampleOnSelectDateFunction,
     },
-    convenienceDates: dateRangeInputConvenienceDates,
   };
 
   // Now add conditional knobs.
@@ -52,6 +58,7 @@ storiesWithKnobs.add('Playground', () => {
       hintMessage: text('Hint Message', ''),
       isDisabled: boolean('isDisabled', false),
       isInvalid: boolean('isInvalid', false),
+      locale: text('Locale', 'en'),
     };
   } else {
     const showLabels = select(
@@ -67,6 +74,7 @@ storiesWithKnobs.add('Playground', () => {
         isGroupLabelHidden: showLabels === 'individual',
         isDisabled: boolean('Group disabled?', false),
         isInvalid: boolean('Group invalid?', false),
+        locale: text('Locale', 'en'),
         validationMessage: text('Validation message for group', ''),
       },
       startDateInputConfig: {
@@ -112,9 +120,19 @@ variations.forEach(variation => {
     delete variationMinusStoryDetails.storyTitle;
 
     const component = isDateRangeInput ? (
-      <XUIDateRangeInputWIP {...variationMinusStoryDetails} />
+      <XUIDateRangeInputWIP
+        locale="en"
+        nextButtonAriaLabel="Next month"
+        prevButtonAriaLabel="Previous month"
+        {...variationMinusStoryDetails}
+      />
     ) : (
-      <XUIDateInputWIP {...variationMinusStoryDetails} />
+      <XUIDateInputWIP
+        locale="en"
+        nextButtonAriaLabel="Next month"
+        prevButtonAriaLabel="Previous month"
+        {...variationMinusStoryDetails}
+      />
     );
 
     return isInFixedContainer ? (